feat: final styling of config + fixes
- final styling and config components split
- temporary edit mode switcher implementation
- few fixes

JIRA: LX-493
risk: low
