+++ libs/sdk-ui-dashboard/styles/scss/visualizationSwitcherToolbar.scss
@@ -2,13 +2,25 @@
 @use "variables";
 @use "@gooddata/sdk-ui-kit/styles/scss/variables" as kit-variables;
 
-.visualization-switcher-toolbar-top {
+.gd-visualization-switcher-toolbar-bubble-wrapper.gd-configuration-bubble-wrapper.overlay-wrapper.sdk-edit-mode-on {
+    .gd-configuration-bubble.gd-visualization-switcher-configuration-bubble {
+        width: 270px;
+    }
+    .bubble-content {
+        padding: 0;
+    }
+    .content {
+        padding: 10px;
+    }
+}
+
+.gd-visualization-switcher-toolbar-top {
     width: 189px;
     height: 40px;
     display: flex;
     justify-content: space-between;
     align-items: center;
-    padding: 6px 8px;
+    padding: 6px 4px 6px 8px;
     margin-bottom: 8px;
     background-color: white;
     .left-section {
@@ -24,7 +36,6 @@
     .middle-section {
         flex: 1 0 auto;
         display: flex;
-        padding: 4px 10px;
         justify-content: space-between;
         align-items: center;
         color: kit-variables.$gd-color-link;
@@ -41,7 +52,7 @@
     }
 
     .right-section {
-        padding: 4px;
+        padding-left: 4px;
         flex: 0 1 auto;
     }
 
@@ -50,11 +61,16 @@
     }
 }
 
-.visualization-switcher-toolbar-bottom {
+.gd-visualization-switcher-toolbar-bottom {
     background-color: white;
 
+    .configuration-panel-header {
+        height: 39px;
+        position: relative;
+    }
+
     .insight-configuration-content {
-        padding: 10px 0;
+        padding: 10px 0 5px 0;
     }
     .no-visualizations-text {
         margin-left: 15px;
@@ -64,32 +80,69 @@
         color: var(--gd-palette-complementary-6, #94a1ad);
     }
 
+    .visualizations-list {
+        max-height: 253px;
+        overflow-y: auto;
+    }
+
     .horizontal-divider {
         width: 100%;
         text-align: center;
-        padding: 0 15px;
+        padding: 5px 15px;
         .horizontal-divider-line {
             height: 1px;
             background-color: variables.$button-grey-background;
         }
     }
     .gd-add-visualization {
-        margin-top: 5px;
         margin-left: 5px;
         font-size: 12px;
     }
 
     .open-visualizations {
-        padding: 15px 0;
+        padding-top: 5px;
+        padding-bottom: 15px;
     }
 
     .switcher-visualizations-list-item {
         display: flex;
         align-items: center;
+        justify-content: space-between;
         padding: 0 15px;
-        &.is-active {
-            .visualization-title {
+        height: 40px;
+        cursor: pointer;
+
+        &:hover {
+            background-color: kit-variables.$is-focused-background;
+
+            .visualization-title,
+            .gd-vis-type {
+                color: kit-variables.$is-focused-color;
+            }
+        }
+
+        &.is-selected {
+            .visualization-title,
+            .gd-vis-type {
+                font-weight: 700;
                 color: var(--gd-palette-primary-0, #14b2e2);
+            }
+            &:hover {
+                background-color: kit-variables.$gd-palette-primary-dimmed;
+            }
+        }
+        .visualization-title {
+            display: flex;
+            align-items: center;
+        }
+        .gd-visualizations-list-item-content-name {
+            white-space: nowrap;
+            max-width: 150px;
+
+            span {
+                display: inline-block;
+                width: 100%;
+                white-space: nowrap;
             }
         }
     }
